DAY01b: Initial commit in PHP 8.0.0
Iterate through the input file line by line.
Add an empty window (=integer array) on every line to a 2d array (=windows).
Push integer to each currently visited window.
The amount of currently visited windows is defined by the window size
itself.
In this case its 3.
If we have more than 3 windows in our windows-array, then we have to
compare the first two windows, conditionally increase a counter and
shift the first window.

This way, we only have to keep 3 windows in memory at a time,
while we stream through the file.